pyproject.toml
[project]
name = "ai-tester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.33.0",
    "ollama>=0.5.3",
    "openai>=1.106.1",
    "pillow>=11.3.0",
]
